Add translation folder and update code
Added a translations folder with english as first file; more to come.
Also changed parts of code to use this file. Tweaked the database for
consistency and stuff. The open/closed status can now be seen from the
owner of a list in the settings. This annoying code at the end of inline
query messages will now be removed automatically after a second. Some
more minor tweaks.