Finish List, Item and User Class
Finished the basic usage of some new classes. This might slow down the
bot a bit more, but makes it way easier to work on it. Have to insert
the new classes into the main py, will do later.
AND this is the first step for efficient notification systems. The plan
is to notify every collaborator after approx. five seconds when someone
made changes. Not like the old bot after every change, to prevent button
spamming and the resulting bot spamming everyone else.